Rearrange the menus for Phabricator
Summary:
Phabricator by default comes with a bunch of stuff that we aren't using at eSpark. I've rearranged the applications to simplify the interface and prioritize those apps that we actually use.

Now, only three apps get full icons: Differential & Wiki, as well as the Feed to view activity (which most people probably aren't aware exists). Below that, there are three other apps that merit attention:
 - Diffusion (for browsing codebase - kind of redundant to Github but useful for invoking audits)
 - Audit (for doing post-review, which we probably don't do enough of)
 - and Herald, for subscribing to specific files, etc -- great for making sure to catch everything even if someone forgets to add you as a reviewer

I also changed the title of Phriction to just Wiki, because what the hell is Phriction, and I removed the Jump Nav because I found it useless.

Hopefully with this simplified model the tool will be more useful. Let me know if you have any suggestions on style.

Test Plan: I configured a local instance of Phabricator and tested the menu here.

Reviewers: mniessner, john

Reviewed By: mniessner

Differential Revision: http://ph.esparklearning.com/D634